Рефакторинг компонентов главной страницы
