+++ project/src/components/offer-item/offer-item.tsx
